refactor: glass effect css & absolute imports - added glassmorphism mixins - remove unused scss vars - add tsconfigpaths vite plugin - update typescript config for absolute paths
