+++ package.json
@@ -24,7 +24,8 @@
 		"lucide-react": "^0.522.0",
 		"react": "^19.1.0",
 		"react-dom": "^19.1.0",
-		"sass": "^1.89.2"
+		"sass": "^1.89.2",
+		"vite-tsconfig-paths": "^5.1.4"
 	},
 	"devDependencies": {
 		"@chromatic-com/storybook": "^4.0.1",
