--- src/components/button/styles.module.scss
@@ -1,30 +1,0 @@
-@use "../../design-system/mixins/base" as *;
-@use "../../design-system/functions/helpers" as *;
-@use "../../design-system/tokens/colors" as *;
-
-.button {
-	@include button-base;
-	display: flex;
-	align-items: center;
-	gap: space(2);
-
-	&.primary {
-		color: color(white);
-		background: var(--color-gradient-primary);
-		border: none;
-
-		&:hover {
-			background: var(--color-gradient-primary-hover);
-		}
-	}
-
-	&.secondary {
-		background: var(--color-background-button);
-		border: 1px solid var(--color-border-button);
-		color: var(--color-text-primary);
-
-		&:hover {
-			background: var(--color-background-button-hover);
-		}
-	}
-}
