+++ src/components/buttons/button/button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import Button from "./";
+import Button from ".";
 
 const meta = {
 	component: Button,
