+++ src/components/inputs/text-input/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./text-input";
