+++ src/components/inputs/text-input/text-input.tsx
@@ -1,0 +1,5 @@
+const TextInput = () => {
+	return <input type="text" />;
+};
+
+export default TextInput;
