+++ src/components/toggle-theme-button/styles.module.scss
@@ -1,19 +1,18 @@
 @use "../../design-system/mixins/base" as *;
+@use "../../design-system/mixins/effects" as *;
 @use "../../design-system/functions/helpers" as *;
 
 .iconButton {
 	@include button-base;
+	@include glass-morphism-background;
 
-	background: var(--color-background-button);
-	border: 1px solid var(--color-border-button);
-	color: var(--color-text-primary);
 	padding: space(3);
 	display: flex;
 	align-items: center;
 	justify-content: center;
 
 	&:hover {
-		background: var(--color-background-button-hover);
+		@include glass-morphism-button-hover;
 	}
 }
