+++ src/design-system/mixins/_effects.scss
@@ -1,0 +1,18 @@
+@use "sass:map";
+@use "../tokens/colors" as *;
+@use "../tokens/effects" as *;
+
+@mixin glass-morphism-background {
+	background: var(--color-background-card);
+	backdrop-filter: blur(1px) saturate(150%);
+	-webkit-backdrop-filter: blur(1px) saturate(150%); // Safari support
+	border: 1px solid var(--color-border-card);
+	color: var(--color-text-primary);
+}
+
+@mixin glass-morphism-button-hover {
+	background: var(--color-background-card-hover);
+	backdrop-filter: blur(10px) saturate(180%);
+	-webkit-backdrop-filter: blur(10px) saturate(180%);
+	border-color: var(--color-border-card-hover);
+}
