+++ src/design-system/themes/_light.scss
@@ -4,36 +4,26 @@
 
 $light-theme: (
 	// Backgrounds
+	// background-primary: linear-gradient(
+	// 		135deg,
+	// 		#f9fafb 0%,
+	// 		#ffffff 50%,
+	// 		#f3f4f6 100%
+	// 	),
 	background-primary: linear-gradient(
-			135deg,
-			#f9fafb 0%,
-			#ffffff 50%,
-			#f3f4f6 100%
-		),
+		135deg,
+		#fdf2f8 0%,
+		#fed7aa 50%,
+		#f9fafb 100%
+	),
 	background-card: color(white-alpha-40),
 	background-card-hover: color(white-alpha-60),
-	background-input: color(white-alpha-60),
-	background-input-focus: color(white-alpha-80),
-	background-summary: color(white-alpha-30),
-	background-add-card: color(white-alpha-30),
-	background-button: color(white-alpha-60),
-	background-button-hover: color(white-alpha-80),
+
 	// Borders
 	border-card: color(gray-200),
-	border-input: color(gray-300),
-	border-input-focus: color(gray-400),
-	border-button: color(gray-200),
-	border-dashed: color(gray-300),
-	border-dashed-hover: color(gray-400),
+	border-card-hover: color(gray-300),
 	// Text
 	text-primary: color(gray-800),
-	text-secondary: color(gray-600),
-	text-muted: color(gray-500),
-	text-label: color(gray-600),
-	text-placeholder: color(gray-500),
-	text-button-secondary: color(gray-500),
-	text-button-secondary-hover: color(gray-700),
-	text-add-card: color(gray-600),
 	// Gradients
 	gradient-primary: map.get($gradients, primary-light),
 	gradient-primary-hover: map.get($gradients, primary-light-hover)
