+++ src/design-system/tokens/_effects.scss
@@ -8,9 +8,10 @@ $effects: (
 	scale-hover: 1.02,
 	scale-card-hover: 1.05,
 
-	// Transitions
-	transition-fast: 200ms ease-in-out,
-	transition-normal: 300ms ease-in-out,
-	transition-slow: 500ms ease-in-out,
-	transition-theme: 700ms ease-in-out
+	// TODO: paused for now
+	// // Transitions
+	// transition-fast: 200ms ease-in-out,
+	// transition-normal: 300ms ease-in-out,
+	// transition-slow: 500ms ease-in-out,
+	// transition-theme: 700ms ease-in-out
 );
