+++ src/features/home/styles.module.scss
@@ -1,7 +1,9 @@
 .home {
 	background: var(--color-background-primary);
-	display: grid;
-	place-items: center;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
 	height: 100vh;
 	width: 100vw;
 }
