+++ tsconfig.app.json
@@ -13,6 +13,15 @@
 		"moduleDetection": "force",
 		"noEmit": true,
 		"jsx": "react-jsx",
+		/* Path mapping */
+		"baseUrl": ".",
+		"paths": {
+			"@/*": ["src/*"],
+			"@/components/*": ["src/components/*"],
+			"@/features/*": ["src/features/*"],
+			"@/context/*": ["src/context/*"],
+			"@/design-system/*": ["src/design-system/*"]
+		},
 		/* Linting */
 		"strict": true,
 		"noUnusedLocals": true,
